drafting big changes
To support formal verification of multi-thread executions it warrants significant
simplification of the compiler to allow more ergonomic implementation at the
exploration stage.

This simplification is making primitive operations non-intrinsic and ultimately
inlining all functions to inline assembly, such that in formal verification the
number of racy operations is reduced to `ldr`, `str` etc.